+++ tests/payment/Payment - Credit card - Expired - vtexgame1nolean.test.js
@@ -1,3 +1,0 @@
-import test from './models/Payment - Credit card - Expired.model.js'
-
-test('vtexgame1nolean')