+++ tests/payment/Payment - Credit card - Redirect - vtexgame1.test.js
@@ -1,3 +1,0 @@
-import test from './models/Payment - Credit card - Redirect.model.js'
-
-test('vtexgame1')