+++ tests/payment/Payment - Google Pay - vtexgame1.test.js
@@ -1,3 +1,0 @@
-import test from './models/Payment - Google Pay.model.js'
-
-test('vtexgame1')