+++ tests/payment/models/Payment - Credit card - Payment App.model.js
@@ -1,4 +1,5 @@
 import { setup, visitAndClearCookies } from '../../../utils'
+import { TIMEOUTS } from '../../../utils/timeouts'
 import {
   fillEmail,
   getRandomEmail,
@@ -36,8 +37,9 @@ export default function test(account) {
         cy.get('#shipping-data').contains('Expressa').should('be.visible')
         cy.get('#shipping-data').contains('PAC Lento').should('be.visible')
       } else {
-        cy.get('#shipping-data').contains('Mais rápida').should('be.visible')
-        cy.get('#shipping-data').contains('Mais econômica').should('be.visible')
+        cy.get('#shipping-data')
+          .contains('Em até 7 dias úteis')
+          .should('be.visible')
       }
 
       goToPayment()
@@ -45,7 +47,10 @@ export default function test(account) {
       completePurchase()
       confirmPaymentApp()
 
-      cy.url({ timeout: 120000 }).should('contain', '/orderPlaced')
+      cy.url({ timeout: TIMEOUTS.PAYMENT_PROCESSING }).should(
+        'contain',
+        '/orderPlaced'
+      )
       cy.wait(2000)
       cy.contains(email).should('be.visible')
       cy.contains('Fernando Application').should('be.visible')
@@ -55,7 +60,7 @@ export default function test(account) {
       cy.contains(DELIVERY_TEXT).should('be.visible')
       cy.contains('Rua Saint Roman 12').should('be.visible')
       cy.contains('Copacabana').should('be.visible')
-      cy.contains('PAC').should('be.visible')
+      cy.contains('Motoboy').should('be.visible')
     })
   })
 }