+++ tests/shipping/models/Delivery_Scheduled Delivery - Second Purchase - Credit card.model.js
@@ -1,4 +1,5 @@
 import { setup, visitAndClearCookies } from '../../../utils'
+import { TIMEOUTS } from '../../../utils/timeouts'
 import {
   fillEmail,
   confirmSecondPurchase,
@@ -31,7 +32,10 @@ export default function test(account) {
       typeCVV()
       completePurchase()
 
-      cy.url({ timeout: 120000 }).should('contain', '/orderPlaced')
+      cy.url({ timeout: TIMEOUTS.PAYMENT_PROCESSING }).should(
+        'contain',
+        '/orderPlaced'
+      )
       cy.wait(2000)
       cy.contains(email).should('be.visible')
       cy.contains(DELIVERY_TEXT).should('be.visible')