+++ tests/shipping/models/Pickup_Delivery - Credit card.model.js
@@ -1,4 +1,5 @@
 import { setup, visitAndClearCookies } from '../../../utils'
+import { TIMEOUTS } from '../../../utils/timeouts'
 import {
   fillEmail,
   getRandomEmail,
@@ -46,15 +47,19 @@ export default function test(account) {
         cy.get('#shipping-data').contains('Expressa').should('be.visible')
         cy.get('#shipping-data').contains('PAC Lento').should('be.visible')
       } else {
-        cy.get('#shipping-data').contains('Mais rápida').should('be.visible')
-        cy.get('#shipping-data').contains('Mais econômica').should('be.visible')
+        cy.get('#shipping-data')
+          .contains('Em até 7 dias úteis')
+          .should('be.visible')
       }
 
       goToPayment()
       payWithCreditCard()
       completePurchase()
 
-      cy.url({ timeout: 120000 }).should('contain', '/orderPlaced')
+      cy.url({ timeout: TIMEOUTS.PAYMENT_PROCESSING }).should(
+        'contain',
+        '/orderPlaced'
+      )
       cy.wait(2000)
       cy.contains(email).should('be.visible')
       cy.contains('Fernando Coelho').should('be.visible')